range enemy, better map generation, player death

+++ Assets/Scripts/Bomb.cs
@@ -31,7 +31,7 @@ public class Bomb : MonoBehaviour,IDamageable
         Destroy(gameObject);
     }
 
-    public void Damage(float damage, Vector3 pos) {
+    public void Damage(int damage, Vector3 pos) {
         Vector3 dir = transform.position - pos;
         dir.Normalize();
         rb.AddForce(dir * damage, ForceMode.Impulse);
