+++ Assets/Scripts/Target.cs
@@ -6,7 +6,7 @@ using DG.Tweening;
 public class Target : MonoBehaviour, IDamageable
 {
     private float health = 100f;
-    public void Damage(float damage, Vector3 pos) {
+    public void Damage(int damage, Vector3 pos) {
         transform.DOPunchScale(0.3f * Vector3.one, 0.3f);
         health -= damage;
         if(health <= 0) {
